show login page any time there is no user_login field in config (meaning user is not authorized)
